pyproject.toml
[tool.poetry]
name = "votecounterer"
version = "0.1.0"
description = ""
authors = ["Martin K <martyshkon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
nextcord = "^2.4.2"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pillow = "^9.5.0"

[tool.poetry.group.dev.dependencies]
taskipy = "^1.11.0"
pre-commit = "^3.3.2"

[tool.poetry.group.lint.dependencies]
isort = "^5.12.0"
black = "^23.3.0"
flake8 = "^6.0.0"

[tool.black]
line_length = 79

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
pre-commit = "pre-commit install"
lint = "pre-commit run --all-files"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
